display OS_CFGINT_TICK_RATE_HZ as unsigned short
git-svn-id: https://micro-os-plus.svn.sourceforge.net/svnroot/micro-os-plus/trunk/Eclipse@1675 ba4b06f3-5e2e-0410-8656-f0fcc57e5dc0

+++ system/portable/kernel/src/OS.cpp
@@ -184,7 +184,7 @@ void OS::earlyInit(void)
 #endif /* defined(OS_EXCLUDE_OSTIMERTICKS_NAKED_ISR) */
 
     OSDeviceDebug::putString_P(PSTR("SysTick: "));
-    OSDeviceDebug::putDec(OS_CFGINT_TICK_RATE_HZ);
+    OSDeviceDebug::putDec((unsigned short)OS_CFGINT_TICK_RATE_HZ);
     OSDeviceDebug::putString_P(PSTR(" ticks/sec"));
     OSDeviceDebug::putNewLine();
 #endif /* defined(DEBUG) && defined(OS_EXCLUDE_OSTIMER) */
